# server/api/openapi.yaml
openapi: 3.0.3
info:
  title: Menguz API
  description: |
    API HTTP de la plataforma **Menguz Vinos & Licores**: tienda en línea (B2C),
    portal mayorista (B2B), panel de administración (JWT), chatbot, sommelier IA
    (SSE) y webhooks de WhatsApp Business (Meta).

    La documentación interactiva Swagger UI se sirve embebida en el propio binario
    en `/docs` y la especificación en `/openapi.yaml`.

    - **Autenticación admin:** `POST /api/auth/login` con correo/contraseña del
      administrador; devuelve `access_token` (JWT HS256, 15 min por defecto) y
      `refresh_token` rotativo de un solo uso (7 días). Envia el access token como
      `Authorization: Bearer <token>` en el grupo `/api/admin/*` (roles `admin` y
      `ventas`).
    - **Autenticación B2B:** `POST /api/b2b/login` con el correo del cliente SAE y
      la contraseña del portal; el token (rol `b2b`) habilita `/api/b2b/*` y precia
      el catálogo con la lista de precios SAE del cliente (1 = menudeo, 2 = medio
      mayorista, 3 = vinatería).
    - **Carrito:** se identifica por la cookie HttpOnly `mgz_cart` (7 días); el
      servidor la crea automáticamente en la primera llamada.
    - **Errores:** las respuestas de error usan el formato `{"message": "..."}`
      (estándar de Echo), con los códigos documentados por operación.
    - **Rate limiting:** límite global por IP en `/api/*`, `/admin/*`, `/data/*`
      y `/webhooks/*`; buckets más estrictos en checkout y chat (429).
...
  version: 0.4.0
...
servers:
...
paths:
...
  /api/products/{cve}/perfil:
    get:
      tags: [Catálogo]
      summary: Ficha de sommelier pública de un producto
      description: |
        Ficha curada (capa editorial) que alimenta al Sommelier IA: uvas, bodega,
        región, notas de cata, maridaje, ocasiones y tags, más precio efectivo,
        existencia e imagen en vivo. Sin autenticación.
      operationId: publicPerfil
      parameters:
        - $ref: '#/components/parameters/CveArt'
      responses:
        '200':
          description: Ficha completa del vino/licor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FichaVino'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /api/admin/products/{cve}/perfil:
    get:
      tags: [Admin]
      summary: Ficha de sommelier de un producto (admin)
      description: Devuelve la ficha curada del producto junto con el producto del catálogo. Requiere JWT con rol `admin` o `ventas`.
      operationId: adminGetPerfil
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CveArt'
      responses:
        '200':
          description: Ficha y producto.
          content:
            application/json:
              schema:
                type: object
                properties:
                  perfil: {$ref: '#/components/schemas/PerfilVino'}
                  producto: {$ref: '#/components/schemas/Producto'}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    put:
      tags: [Admin]
      summary: Guardar ficha de sommelier
      description: |
        Crea o actualiza (upsert) la ficha curada del producto; el cuerpo es un
        `PerfilVino` completo (`cve_art` se toma de la ruta). La ficha alimenta
        al Sommelier IA y se conserva entre sincronizaciones SAE.
        Requiere JWT con rol `admin` o `ventas`.
      operationId: adminSavePerfil
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CveArt'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PerfilVino'}
      responses:
        '200':
          description: Ficha guardada.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PerfilVino'}
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/admin/perfiles/bootstrap:
    post:
      tags: [Admin]
      summary: Generar fichas borrador
      description: |
        Genera fichas de sommelier borrador (tipo y uvas detectadas en el nombre)
        para los productos sin ficha. No pisa las fichas editadas a mano.
        Requiere JWT con rol `admin` o `ventas`.
      operationId: adminBootstrapPerfiles
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Resultado del generador.
          content:
            application/json:
              schema:
                type: object
                properties:
                  creados: {type: integer, example: 23}
                  actualizados: {type: integer, example: 4}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT HS256 firmado con `JWT_SECRET`. Claims: `sub` (ID de usuario o
        cliente), `email`, `rol` (`admin` | `ventas` | `b2b`), `lista` (lista de
        precios SAE, solo B2B), `iss=menguz`. Access token 15 min (configurable);
        rota con el refresh token de un solo uso vía `POST /api/auth/refresh`.
  parameters:
...
    CveArt:
      name: cve
      in: path
      required: true
      description: Clave de artículo SAE.
      schema: {type: string, example: VIN-TIN-001}
...
  responses:
    BadRequest:
      description: Petición inválida (JSON mal formado, validación o regla de negocio).
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example: {message: JSON inválido}
    Unauthorized:
      description: 'Falta el header `Authorization: Bearer` o el token es inválido/expirado.'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          examples:
            sinToken: {value: {message: token requerido}}
            tokenInvalido: {value: {message: token inválido o expirado}}
    Forbidden:
      description: Token válido pero rol no autorizado para este grupo.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example: {message: rol no autorizado}
    NotFound:
      description: Recurso no encontrado.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example: {message: producto no encontrado}
    TooManyRequests:
      description: Límite de solicitudes excedido (rate limit por IP); reintentar más tarde.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example: {message: demasiadas solicitudes, intenta más tarde}
  schemas:
    Error:
      type: object
      properties:
        message: {type: string}
    Producto:
      type: object
      description: Producto sincronizado desde Aspel SAE 8.0 (INVE01 + INVE_CLIB01).
      properties:
        cve_art: {type: string, example: VIN-TIN-001}
        nombre: {type: string, example: Malbec Reserva 2021}
        descripcion: {type: string, example: Malbec de Valle de Guadalupe, 12 meses en roble.}
        precio_base: {type: number, example: 489.0, description: Precio lista 1 (menudeo).}
        precio_lista_2: {type: number, example: 439.0, description: Precio medio mayorista.}
        precio_lista_3: {type: number, example: 399.0, description: Precio vinatería / B2B.}
        precio_oferta: {type: number, example: 429.0, description: Override de precio desde campaña (omite si es 0).}
        existencia: {type: number, example: 12.0}
        unidad_medida: {type: string, example: PZA}
        linea: {type: string, example: VINOS}
        categoria: {type: string, example: Vino Tinto, description: Categoría de la tienda.}
        grupo: {type: string, example: Vinos, description: Grupo de la tienda (Vinos, Licores…).}
        graduacion: {type: string, example: 13.5%}
        origen: {type: string, example: México}
        notas_cata: {type: string, example: 'Color rubí intenso; nariz a frutos rojos y vainilla.'}
        imagen_url: {type: string, example: /products/malbec-reserva.png}
        activo: {type: boolean, example: true}
        ultima_sync: {type: string, format: date-time}
...
    PerfilVino:
      type: object
      description: Ficha curada de sommelier (capa editorial sobre el producto SAE); se conserva entre syncs.
      properties:
        cve_art: {type: string, example: VIN-TIN-001}
        tipo: {type: string, example: Tinto, enum: [Tinto, Blanco, Rosado, Espumoso, Champagne, Licor, Otro]}
        uvas: {type: array, items: {type: string}, example: [Malbec, Cabernet Sauvignon]}
        bodega: {type: string, example: Viñedos del Valle}
        region: {type: string, example: Valle de Guadalupe}
        pais: {type: string, example: México}
        anada: {type: string, example: '2021', description: Añada; vacío = sin añada (NV).}
        graduacion: {type: string, example: 13.5%}
        crianza: {type: string, example: Roble 6 meses}
        notas_cata: {type: string, example: 'Color rubí; nariz a frutos rojos; boca tánica elegante.'}
        maridaje: {type: string, example: Carnes rojas a la parrilla, quesos añejos}
        descripcion: {type: string, example: Descripción editorial larga del vino.}
        ocasiones: {type: array, items: {type: string}, example: [asado, regalo, celebración]}
        tags: {type: array, items: {type: string}, example: [frutal, tánico, añejo]}
        updated_at: {type: string, format: date-time}
    FichaVino:
      type: object
      description: Ficha pública completa (identidad + precio efectivo + existencia + perfil curado).
      properties:
        cve_art: {type: string, example: VIN-TIN-001}
        nombre: {type: string, example: Malbec Reserva 2021}
        categoria: {type: string, example: Vino Tinto}
        tipo: {type: string, example: Tinto}
        precio: {type: number, example: 489.0, description: Precio efectivo (oferta aplicada si aplica).}
        existencia: {type: number, example: 12.0}
        imagen: {type: string, example: products/malbec-reserva.png}
        descripcion: {type: string, example: Descripción editorial o la del catálogo SAE.}
        bodega: {type: string, example: Viñedos del Valle}
        uvas: {type: array, items: {type: string}, example: [Malbec]}
        region: {type: string, example: Valle de Guadalupe}
        pais: {type: string, example: México}
        anada: {type: string, example: '2021', description: '`sin añada (NV)` cuando no aplica.'}
        graduacion: {type: string, example: 13.5%}
        crianza: {type: string, example: Roble 6 meses}
        notas_cata: {type: string, example: 'Color rubí; nariz a frutos rojos.'}
        maridaje: {type: string, example: Carnes rojas a la parrilla}
        ocasiones: {type: array, items: {type: string}, example: [asado]}
        tags: {type: array, items: {type: string}, example: [frutal, tánico]}
...